requirements.txt
attrs==23.2.0
bcrypt==4.1.2
blinker==1.7.0
cachetools==5.3.2
certifi==2024.2.2
chardet==5.2.0
charset-normalizer==3.3.2
click==8.1.7
cloudpickle==3.0.0
colorlog==6.8.2
contourpy==1.2.1
cycler==0.12.1
dask==2024.6.0
dominate==2.9.1
et-xmlfile==1.1.0
extra-streamlit-components==0.1.71
fastjsonschema==2.20.0
fonttools==4.53.0
fsspec==2024.6.0
gitdb==4.0.11
GitPython==3.1.41
idna==3.6
importlib_resources==6.4.0
ipynbname==2024.1.0.0
ipython-genutils==0.2.0
isoweek==1.3.3
Jinja2==3.1.3
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
kiwisolver==1.4.5
lightgbm==4.4.0
locket==1.0.0
lxml==4.9.4
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.9.0
mdurl==0.1.2
micawber==0.5.5
multimethod==1.11.2
nbformat==5.10.4
numpy==1.26.4
openpyxl==3.1.2
pandas==2.2.2
partd==1.4.2
pillow==10.2.0
plotly==5.18.0
kaleido==0.2.1
protobuf==4.25.2
pyarrow==15.0.0
pydeck==0.8.1b0
pydub==0.25.1
PyJWT==2.8.0
pyparsing==3.1.2
PyPDF2==3.0.1
pytz==2024.1
PyYAML==6.0.1
referencing==0.33.0
requests==2.31.0
rich==13.7.0
rpds-py==0.17.1
scipy==1.13.1
seaborn==0.13.2
smmap==5.0.1
streamlit==1.31.0
streamlit-authenticator==0.3.2
tenacity==8.2.3
toml==0.10.2
toolz==0.12.1
typing_extensions==4.9.0
tzdata==2023.4
tzlocal==5.2
urllib3==2.2.0
validators==0.22.0
vega-datasets==0.9.0
xgboost==2.0.3
xlrd==2.0.1
joblib==1.4.2
scikit-learn==1.6.1
nltk==3.9.1
python-pptx==1.0.2
markdown-it-py==3.0.0
markdown
beautifulsoup4
openai>=1.0.0
anthropic>=0.18.0